compatibility page
- added first shades on the compatibility page
